// src/framework/shopify/schemas/schema.d.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
    [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
    [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
    [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    /** An ISO-8601 encoded UTC date time string. Example value: `"2019-07-03T20:47:55Z"`. */
    DateTime: any;
    /** A signed decimal number, which supports arbitrary precision and is serialized as a string. Example value: `"29.99"`. */
    Decimal: any;
    /** A string containing HTML code. Example value: `"<p>Grey cotton knit sweater.</p>"`. */
    HTML: any;
    /** A monetary value string. Example value: `"100.57"`. */
    Money: any;
    /**
     * An RFC 3986 and RFC 3987 compliant URI string.
     *
     * Example value: `"https://johns-apparel.myshopify.com"`.
     *
     */
    URL: any;
};
// ...
export type CollectionConnection = {
    __typename?: "CollectionConnection";
    /** A list of edges. */
    edges: Array<CollectionEdge>;
    /** Information to aid in pagination. */
    pageInfo: PageInfo;
};

/** An auto-generated type which holds one Collection and a cursor during pagination. */
export type CollectionEdge = {
    __typename?: "CollectionEdge";
    /** A cursor for use in pagination. */
    cursor: Scalars["String"];
    /** The item at the end of CollectionEdge. */
    node: Collection;
};
// ...
export type HasMetafields = {
    /** The metafield associated with the resource. */
    metafield?: Maybe<Metafield>;
    /** A paginated list of metafields associated with the resource. */
    metafields: MetafieldConnection;
};
// ...
export type Image = {
    __typename?: "Image";
    /** A word or phrase to share the nature or contents of an image. */
    altText?: Maybe<Scalars["String"]>;
    /** The original height of the image in pixels. Returns `null` if the image is not hosted by Shopify. */
    height?: Maybe<Scalars["Int"]>;
    /** A unique identifier for the image. */
    id?: Maybe<Scalars["ID"]>;
    /**
     * The location of the original image as a URL.
     *
     * If there are any existing transformations in the original source URL, they will remain and not be stripped.
     */
    originalSrc: Scalars["URL"];
    /**
     * The location of the image as a URL.
     * @deprecated Previously an image had a single `src` field. This could either return the original image
     * location or a URL that contained transformations such as sizing or scale.
     *
     * These transformations were specified by arguments on the parent field.
     *
     * Now an image has two distinct URL fields: `originalSrc` and `transformedSrc`.
     *
     * * `originalSrc` - the original unmodified image URL
     * * `transformedSrc` - the image URL with the specified transformations included
     *
     * To migrate to the new fields, image transformations should be moved from the parent field to `transformedSrc`.
     *
     * Before:
     * ```graphql
     * {
     *   shop {
     *     productImages(maxWidth: 200, scale: 2) {
     *       edges {
     *         node {
     *           src
     *         }
     *       }
     *     }
     *   }
     * }
     * ```
     *
     * After:
     * ```graphql
     * {
     *   shop {
     *     productImages {
     *       edges {
     *         node {
     *           transformedSrc(maxWidth: 200, scale: 2)
     *         }
     *       }
     *     }
     *   }
     * }
     * ```
     *
     */
    src: Scalars["URL"];
    /**
     * The location of the transformed image as a URL.
     *
     * All transformation arguments are considered "best-effort". If they can be applied to an image, they will be.
     * Otherwise any transformations which an image type does not support will be ignored.
     */
    transformedSrc: Scalars["URL"];
    /** The original width of the image in pixels. Returns `null` if the image is not hosted by Shopify. */
    width?: Maybe<Scalars["Int"]>;
};

/** Represents an image resource. */
export type ImageTransformedSrcArgs = {
    maxWidth?: Maybe<Scalars["Int"]>;
    maxHeight?: Maybe<Scalars["Int"]>;
    crop?: Maybe<CropRegion>;
    scale?: Maybe<Scalars["Int"]>;
    preferredContentType?: Maybe<ImageContentType>;
};

/** An auto-generated type for paginating through multiple Images. */
export type ImageConnection = {
    __typename?: "ImageConnection";
    /** A list of edges. */
    edges: Array<ImageEdge>;
    /** Information to aid in pagination. */
    pageInfo: PageInfo;
};

/** List of supported image content types. */
export enum ImageContentType {
    /** A PNG image. */
    Png = "PNG",
    /** A JPG image. */
    Jpg = "JPG",
    /** A WEBP image. */
    Webp = "WEBP",
}

/** An auto-generated type which holds one Image and a cursor during pagination. */
export type ImageEdge = {
    __typename?: "ImageEdge";
    /** A cursor for use in pagination. */
    cursor: Scalars["String"];
    /** The item at the end of ImageEdge. */
    node: Image;
};
// ...
export type MediaConnection = {
    __typename?: "MediaConnection";
    /** A list of edges. */
    edges: Array<MediaEdge>;
    /** Information to aid in pagination. */
    pageInfo: PageInfo;
};
// ...
export type MediaEdge = {
    __typename?: "MediaEdge";
    /** A cursor for use in pagination. */
    cursor: Scalars["String"];
    /** The item at the end of MediaEdge. */
    node: Media;
};
// ...
export type Metafield = Node & {
    __typename?: "Metafield";
    /** The date and time when the storefront metafield was created. */
    createdAt: Scalars["DateTime"];
    /** The description of a metafield. */
    description?: Maybe<Scalars["String"]>;
    /** Globally unique identifier. */
    id: Scalars["ID"];
    /** The key name for a metafield. */
    key: Scalars["String"];
    /** The namespace for a metafield. */
    namespace: Scalars["String"];
    /** The parent object that the metafield belongs to. */
    parentResource: MetafieldParentResource;
    /** The date and time when the storefront metafield was updated. */
    updatedAt: Scalars["DateTime"];
    /** The value of a metafield. */
    value: Scalars["String"];
    /** Represents the metafield value type. */
    valueType: MetafieldValueType;
};

/** An auto-generated type for paginating through multiple Metafields. */
export type MetafieldConnection = {
    __typename?: "MetafieldConnection";
    /** A list of edges. */
    edges: Array<MetafieldEdge>;
    /** Information to aid in pagination. */
    pageInfo: PageInfo;
};

/** An auto-generated type which holds one Metafield and a cursor during pagination. */
export type MetafieldEdge = {
    __typename?: "MetafieldEdge";
    /** A cursor for use in pagination. */
    cursor: Scalars["String"];
    /** The item at the end of MetafieldEdge. */
    node: Metafield;
};

/** A resource that the metafield belongs to. */
export type MetafieldParentResource = Product | ProductVariant;

/** Metafield value types. */
export enum MetafieldValueType {
    /** A string metafield. */
    String = "STRING",
    /** An integer metafield. */
    Integer = "INTEGER",
    /** A json string metafield. */
    JsonString = "JSON_STRING",
}
// ...
export type MoneyV2 = {
    __typename?: "MoneyV2";
    /** Decimal money amount. */
    amount: Scalars["Decimal"];
    /** Currency of the money. */
    currencyCode: CurrencyCode;
};
// ...
export type Node = {
    /** Globally unique identifier. */
    id: Scalars["ID"];
};
// ...
export type PageInfo = {
    __typename?: "PageInfo";
    /** Indicates if there are more pages to fetch. */
    hasNextPage: Scalars["Boolean"];
    /** Indicates if there are any pages prior to the current page. */
    hasPreviousPage: Scalars["Boolean"];
};
// ...
export type Product = Node &
    HasMetafields & {
        __typename?: "Product";
        /** Indicates if at least one product variant is available for sale. */
        availableForSale: Scalars["Boolean"];
        /** List of collections a product belongs to. */
        collections: CollectionConnection;
        /** The compare at price of the product across all variants. */
        compareAtPriceRange: ProductPriceRange;
        /** The date and time when the product was created. */
        createdAt: Scalars["DateTime"];
        /** Stripped description of the product, single line with HTML tags removed. */
        description: Scalars["String"];
        /** The description of the product, complete with HTML formatting. */
        descriptionHtml: Scalars["HTML"];
        /**
         * A human-friendly unique string for the Product automatically generated from its title.
         * They are used by the Liquid templating language to refer to objects.
         */
        handle: Scalars["String"];
        /** Globally unique identifier. */
        id: Scalars["ID"];
        /** List of images associated with the product. */
        images: ImageConnection;
        /** The media associated with the product. */
        media: MediaConnection;
        /** The metafield associated with the resource. */
        metafield?: Maybe<Metafield>;
        /** A paginated list of metafields associated with the resource. */
        metafields: MetafieldConnection;
        /**
         * The online store URL for the product.
         * A value of `null` indicates that the product is not published to the Online Store sales channel.
         */
        onlineStoreUrl?: Maybe<Scalars["URL"]>;
        /** List of product options. */
        options: Array<ProductOption>;
        /** List of price ranges in the presentment currencies for this shop. */
        presentmentPriceRanges: ProductPriceRangeConnection;
        /** The price range. */
        priceRange: ProductPriceRange;
        /** A categorization that a product can be tagged with, commonly used for filtering and searching. */
        productType: Scalars["String"];
        /** The date and time when the product was published to the channel. */
        publishedAt: Scalars["DateTime"];
        /** The product's SEO information. */
        seo: Seo;
        /**
         * A comma separated list of tags that have been added to the product.
         * Additional access scope required for private apps: unauthenticated_read_product_tags.
         */
        tags: Array<Scalars["String"]>;
        /** The product’s title. */
        title: Scalars["String"];
        /** The total quantity of inventory in stock for this Product. */
        totalInventory?: Maybe<Scalars["Int"]>;
        /**
         * The date and time when the product was last modified.
         * A product's `updatedAt` value can change for different reasons. For example, if an order
         * is placed for a product that has inventory tracking set up, then the inventory adjustment
         * is counted as an update.
         */
        updatedAt: Scalars["DateTime"];
        /**
         * Find a product’s variant based on its selected options.
         * This is useful for converting a user’s selection of product options into a single matching variant.
         * If there is not a variant for the selected options, `null` will be returned.
         */
        variantBySelectedOptions?: Maybe<ProductVariant>;
        /** List of the product’s variants. */
        variants: ProductVariantConnection;
        /** The product’s vendor name. */
        vendor: Scalars["String"];
    };
// ...
export type ProductOption = Node & {
    __typename?: "ProductOption";
    /** Globally unique identifier. */
    id: Scalars["ID"];
    /** The product option’s name. */
    name: Scalars["String"];
    /** The corresponding value to the product option name. */
    values: Array<Scalars["String"]>;
};

/** The price range of the product. */
export type ProductPriceRange = {
    __typename?: "ProductPriceRange";
    /** The highest variant's price. */
    maxVariantPrice: MoneyV2;
    /** The lowest variant's price. */
    minVariantPrice: MoneyV2;
};

/** An auto-generated type for paginating through multiple ProductPriceRanges. */
export type ProductPriceRangeConnection = {
    __typename?: "ProductPriceRangeConnection";
    /** A list of edges. */
    edges: Array<ProductPriceRangeEdge>;
    /** Information to aid in pagination. */
    pageInfo: PageInfo;
};

/** An auto-generated type which holds one ProductPriceRange and a cursor during pagination. */
export type ProductPriceRangeEdge = {
    __typename?: "ProductPriceRangeEdge";
    /** A cursor for use in pagination. */
    cursor: Scalars["String"];
    /** The item at the end of ProductPriceRangeEdge. */
    node: ProductPriceRange;
};
// ...
export type ProductVariantConnection = {
    __typename?: "ProductVariantConnection";
    /** A list of edges. */
    edges: Array<ProductVariantEdge>;
    /** Information to aid in pagination. */
    pageInfo: PageInfo;
};

/** An auto-generated type which holds one ProductVariant and a cursor during pagination. */
export type ProductVariantEdge = {
    __typename?: "ProductVariantEdge";
    /** A cursor for use in pagination. */
    cursor: Scalars["String"];
    /** The item at the end of ProductVariantEdge. */
    node: ProductVariant;
};
// ...
export type Seo = {
    __typename?: "SEO";
    /** The meta description. */
    description?: Maybe<Scalars["String"]>;
    /** The SEO title. */
    title?: Maybe<Scalars["String"]>;
};